moved def process_guess_var() and get_phase_mission_bus_length() into mission.utils.py
