+++ aviary/core/AviaryGroup.py
@@ -29,7 +29,7 @@ from aviary.subsystems.mass.mass_builder import CoreMassBuilder
 from aviary.subsystems.premission import CorePreMission
 from aviary.subsystems.propulsion.propulsion_builder import CorePropulsionBuilder
 from aviary.utils.functions import set_warning_format
-from aviary.utils.utils import get_phase_mission_bus_lengths, process_guess_var
+from aviary.mission.utils import get_phase_mission_bus_lengths, process_guess_var
 from aviary.variable_info.variables import Aircraft, Dynamic, Mission, Settings
 from aviary.variable_info.variable_meta_data import _MetaData as BaseMetaData
 from aviary.mission.gasp_based.phases.time_integration_traj import FlexibleTraj
